Revert loss head PR (#2873)
