+++ ultralytics/yolo/engine/trainer.py
@@ -325,7 +325,8 @@ class BaseTrainer:
                 # Forward
                 with torch.cuda.amp.autocast(self.amp):
                     batch = self.preprocess_batch(batch)
-                    self.loss, self.loss_items = de_parallel(self.model).loss(batch)
+                    preds = self.model(batch['img'])
+                    self.loss, self.loss_items = self.criterion(preds, batch)
                     if RANK != -1:
                         self.loss *= world_size
                     self.tloss = (self.tloss * i + self.loss_items) / (i + 1) if self.tloss is not None \
@@ -494,6 +495,12 @@ class BaseTrainer:
     def build_dataset(self, img_path, mode='train', batch=None):
         """Build dataset"""
         raise NotImplementedError('build_dataset function not implemented in trainer')
+
+    def criterion(self, preds, batch):
+        """
+        Returns loss and individual loss items as Tensor.
+        """
+        raise NotImplementedError('criterion function not implemented in trainer')
 
     def label_loss_items(self, loss_items=None, prefix='train'):
         """
